+++ index.html
@@ -228,6 +228,6 @@
   </template>
 
 </body>
-<script src="js/functions.js"></script>
-<script src="js/main.js"></script>
+<!-- <script src="js/functions.js"></script> -->
+<script type="module" src="js/main.js"></script>
 </html>
